Just about works, but still lots of tidying up.  Next is to work out how to dynamically create an instance at config of the correct type
