+++ custom_components/lednetwf_ble/const.py
@@ -13,125 +13,6 @@ STRIP_LIGHT_MODEL = 0x56
 
 # The names given to the effects are just what I thought they looked like.  Updates are welcome.
 
-EFFECTS_LIST_0x53 = [
-    "Gold Ring",
-    "Red Magenta Fade",
-    "Yellow Magenta Fade",
-    "Green Yellow Fade",
-    "Green Blue Spin",
-    "Blue Spin",
-    "Purple Pink Spin",
-    "Color Fade",
-    "Red Blue Flash",
-    "CMRGB Spin",
-    "RGBYMC Follow",
-    "CMYRGB Spin",
-    "RGB Chase",
-    "RGB Tri Reverse Spin",
-    "Red Fade",
-    "Blue Yellow Quad Static",
-    "Red Green Quad Static",
-    "Cyan Magenta Quad Static",
-    "Red Green Reverse Chase",
-    "Blue Yellow Reverse Chase",
-    "Cyan Magenta Reverse Chase",
-    "Yellow RGB Reverse Spin",
-    "Cyan RGB Reverse Spin",
-    "Magenta RGB Reverse Spin",
-    "RGB Reverse Spin",
-    "RGBY Reverse Spin",
-    "Magenta RGBY Reverse Spin",
-    "Cyan RGBYMC Reverse Spin",
-    "White RGBYMC Reverse Spin",
-    "Red Green Reverse Chase 2",
-    "Blue Yellow Reverse Chase 2",
-    "Cyan Pink Reverse Chase",
-    "White Strobe",
-    "White Strobe 2",
-    "Warm White Strobe",
-    "Smooth Color Fade",
-    "White Static",
-    "Pinks Fade",
-    "Cyans Fade",
-    "Cyan Magenta Slow Fade",
-    "Green Yellow Fade 2",
-    "RGBCMY Slow Fade",
-    "Whites Fade",
-    "Pink Purple Fade",
-    "Cyan Magenta Fade",
-    "Cyan Blue Fade",
-    "Yellow Cyan Fade",
-    "Red Yellow Fade",
-    "RGBCMY Strobe",
-    "Warm Cool White Strobe",
-    "Magenta Strobe",
-    "Cyan Strobe",
-    "Yellow Strobe",
-    "Magenta Cyan Strobe",
-    "Cyan Yellow Strobe",
-    "Cool White Strobe Random",
-    "Warm White Strobe Random",
-    "Light Green Strobe Random",
-    "Magenta Strobe Random",
-    "Cyan Strobe Random",
-    "Oranges Ring",
-    "Blue Ring",
-    "RMBCGY Loop",
-    "Cyan Magenta Follow",
-    "Yellow Green Follow",
-    "Pink Blue Follow",
-    "BGP Pastels Loop",
-    "CYM Follow",
-    "Pink Purple Demi Spinner",
-    "Blue Pink Spinner",
-    "Green Spinner",
-    "Blue Yellow Tri Spinner",
-    "Red Yellow Tri Spinner",
-    "Pink Green Tri Spinner",
-    "Red Blue Demi Spinner",
-    "Yellow Green Demi Spinner",
-    "RGB Tri Spinner",
-    "Red Magenta Demi Spinner",
-    "Cyan Magenta Demi Spinner",
-    "RCBM Quad Spinner",
-    "RGBCMY Spinner",
-    "RGB Spinner",
-    "CMB Spinner",
-    "Red Blue Demi Spinner",
-    "Cyan Magenta Demi Spinner",
-    "Yellow Orange Demi Spinner",
-    "Red Blue Striped Spinner",
-    "Green Yellow Striped Spinner",
-    "Red Pink Yellow Striped Spinner",
-    "Cyan Blue Magenta Striped Spinner",
-    "Pastels Striped Spinner",
-    "Rainbow Spin",
-    "Red Pink Blue Spinner",
-    "Cyan Magenta Spinner",
-    "Green Cyan Spinner",
-    "Yellow Red Spinner",
-    "Rainbow Strobe",
-    "Magenta Strobe",
-    "Yellow Orange Demi Strobe",
-    "Yellow Cyan Demi Flash",
-    "White Lightening Strobe",
-    "Purple Lightening Strobe",
-    "Magenta Lightening Strobe",
-    "Yellow Lightening Strobe",
-    "Blue With Sparkles",
-    "Red With Sparkles",
-    "Blue With Sparkles",
-    "Yellow Dissolve",
-    "Magenta Dissolve",
-    "Cyan Dissolve",
-    "Red Green Dissolve",
-    "RGB Dissolve",
-    "RGBCYM Dissolve",
-    "Nothing",
-    "Nothing 2",
-    "_Cycle Through All Modes"
-]
-
 # 0x56 Effect data
 EFFECT_MAP_0x56 = {}
 for e in range(1,100):
